Add Compress_apk property to android_app_import
This change adds a new boolean property, Compress_apk, to
the android_app_import module type. When set to true, the output APK
will be compressed using gzip. This can reduce the size of the
installed APK, saving storage space.

This functionality is the same as the LOCAL_COMPRESSED_MODULE variable
in Android.mk builds.

Bug: 384848192
Bug: 383642702
Test: add compress_apk in android_app_import
Change-Id: If6793ea09951c54d42ba5cc4249b698b676cbcc7
Signed-off-by: texascake <tiktodz@soulvibe-team.cloud>